Sequential design probabilities

Computes sequential probabilities, for each stage of a given one-sided sequential design with specified critical limits and optional futility stops (stopping bounds). When applying the design in practice, at each stage, all data collected up to this point is used to calculate a combined test statistic. Then, if the test statistic is larger than the critical limit at the specific stage, the study is stopped and the null hypothesis rejected. If futility stops have been specified and the test statistic falls below them at any stage, the study is also stopped due to futility while the null hypothesis in this case cannot be discarded.

seqDesign(
  crit,
  mu = 0,
  sd = 1,
  tau = seq_along(crit)/length(crit),
  futStop = rep(-Inf, length(crit)),
  n = 1,
  ...
)

Arguments

crit
numeric vector of critical limits for stopping the study with rejection of the null hypothesis. The length of 'crit' determines the number of stages of the design.

mu
numeric the mean (e.g. mu == 0 under H0). Under H1 this is also known as the drift parameter.

sd
numeric standard deviation (or vector of sd)

tau
numeric vector of monotonely increasing information times.

futStop
numeric vector of futility stopping bounds (optional).

n
numeric sample size, which only has an effect if mu != 0. Useful to compute expected sample size under H1.

...
further arguments passed to seqProb

Value

result is a list with the following components:

exitP the alpha spent at each stage. The cumulative sum of this will provides the cumulative exit probability up to each stage, that is, the probability of having stopped due to exceeding a bound.

undecided probability of continuing the study. In case of the last stage, this means stopping without declaring significance.

cumFutStop cumulative futility stop probabilities.

nk planned sample size per stage.

kExp expected sample size per stage - the sum over all values will provide the total expected sample size.

Examples

# 1-stage
# #######
alpha <- 0.05
crit <- qnorm(1-alpha)
seqDesign(crit)$p$exit                          # type1: 0.05 == alpha

#> [1] 0.05

seqDesign(crit, mu=0.3)$p$exit                  # power: 9%

#> [1] 0.08933625

# Increase sample size
seqDesign(crit, mu=0.3, n=69)$p$exit            # power: 80.1%

#> [1] 0.8015397

# 2-stage
# #######
crit2 <- rep(crit, 2)

# Using two stages with same critical bounds will exceed alpha
cumsum(seqDesign(crit2, mu=0)$p$exit)           # cumulative type1: > 0.08

#> [1] 0.05000000 0.08007553

cumsum(seqDesign(crit2, mu=0, n=69)$p$exit)     # type1 does not depend on n

#> [1] 0.05000000 0.08007553

# Try Bonferroni:
bonf <- rep(qnorm(1-alpha/2), 2)
cumsum(seqDesign(bonf)$p$exit)                  # cumulative type1: 0.0415...

#> [1] 0.02500000 0.04155891

sum(seqDesign(bonf, mu=0.3, n=69)$p$exit)       # cumulative power: 73.1%

#> [1] 0.7313453

# Bonferroni did not use all alpha, so there is room to improve power, for
# example using Pocock like boundaries:
pocock <- c(1.8662, 1.8848)
sum(seqDesign(pocock)$p$exit)                   # cumulative type1: 0.05

#> [1] 0.0500047

sum(seqDesign(pocock, mu=.3, n=69)$p$exit)      # power: 75.8%

#> [1] 0.7575743

# Still the one stage power will never be reached, which is the price to pay
# for the two potential tests. On the other hand, the expected sample size is
# considerably smaller in comparison with the 1-stage design.
sum(seqDesign(pocock, mu=.3, n=69)$n$nkExp)     # expected sample size: 53

#> [1] 53.18018

# Increasing max sample size to 98, yields expected n = 69 and power 87.8%
res <- seqDesign(pocock, mu=.3, n=98)
lapply(unlist(res, recursive=FALSE)[c("n.nkExp", "p.exit")], sum)

#> $n.nkExp
#> [1] 68.97094
#> 
#> $p.exit
#> [1] 0.8779636
#>