Sequential probabilities

Compute sequential probabilities based on normally distributed test statistics and assuming independent increments.

seqProb(
  up,
  lo = rep(-Inf, length(up)),
  tau = seq_along(up)/length(up),
  mean = 0,
  sd = 1/sqrt(tau),
  hasZbounds = TRUE,
  exitLast = FALSE,
  alg = mvtnorm::Miwa()
)

Arguments

up
numeric upper bounds

lo
numeric lower bounds

tau
numeric vector of monotonely increasing information times

mean
numeric the mean

sd
numeric vector of standard deviations

hasZbounds
logical if true, bounds are assumed to be standardized already, otherwise they will be standardized by the function.

exitLast
logical if TRUE, upper bound at the last stage is set to Inf, and lower bound to user-specified upper bound, which means to compute the probability to stay below the upper bounds in all stages except the last stage. In a sequential study design this corresponds to accepting the null hypothesis in stages 1 to n-1 but rejecting it at stage n, i.e., the probability to exit the study just at stage n.

alg
algorithm object to be used for numerical integration (see mvtnorm::pmvnorm).

Value

numeric calculated (multivariate) normal probability to stay between bounds, or (if exitLast was set to TRUE) to stay between bounds and exceed upper bound at last stage.

See also

Examples

# 1-D
pnorm(0)               == seqProb(0)                                   # TRUE

#> Error in seqProb(0): could not find function "seqProb"

pnorm(1, mean=1)       == seqProb(1, mean=1)                           # TRUE

#> Error in seqProb(1, mean = 1): could not find function "seqProb"

pnorm(2, mean=1, sd=2) == seqProb(2, mean=1, sd=2, hasZbounds=FALSE)   # TRUE

#> Error in seqProb(2, mean = 1, sd = 2, hasZbounds = FALSE): could not find function "seqProb"

seqProb(-Inf)                                        # 0

#> Error in seqProb(-Inf): could not find function "seqProb"

seqProb(Inf)                                         # 1

#> Error in seqProb(Inf): could not find function "seqProb"

seqProb(qnorm(.95))                                  # .95

#> Error in seqProb(qnorm(0.95)): could not find function "seqProb"

seqProb(qnorm(.95), exitLast=TRUE)                   # 1-.95

#> Error in seqProb(qnorm(0.95), exitLast = TRUE): could not find function "seqProb"

# Multivariate
seqProb(c(Inf, 0)) == seqProb(0)                       # TRUE

#> Error in seqProb(c(Inf, 0)): could not find function "seqProb"

p <- seqProb(c(2, 1), mean=0)

#> Error in seqProb(c(2, 1), mean = 0): could not find function "seqProb"

seqProb(c(2, 1), mean=0, tau=c(.25, 0.5)) == p         # TRUE

#> Error in seqProb(c(2, 1), mean = 0, tau = c(0.25, 0.5)): could not find function "seqProb"

seqProb(c(2, 1), mean=0, tau=c(.25, 1)) == p           # FALSE

#> Error in seqProb(c(2, 1), mean = 0, tau = c(0.25, 1)): could not find function "seqProb"